order is now 32 bits instead of 64 bits

+++ streamconnector/workflowserializerv1.cpp
@@ -79,8 +79,8 @@ bool WorkflowSerializerV1::storeNode(const SPWorkFlowNode& node, const IOOptions
         _stream << sz;
         for(qint32 t=0; t < sz; ++t){
             WorkFlowCondition::Test test = condition->test(t)    ;
-            _stream << (int)test._pre;
-            _stream << (int)test._post;
+            _stream << (qint32)test._pre;
+            _stream << (qint32)test._post;
            storeNode(test._operation);
         }
         auto operations = condition->subnodes("operations")    ;
@@ -91,7 +91,7 @@ bool WorkflowSerializerV1::storeNode(const SPWorkFlowNode& node, const IOOptions
     }
     node->box().store(_stream);
 
-    int count = node->inputCount();
+    qint32 count = node->inputCount();
     if ( node->type() == "junctionnode")
         // junctions have 3 parameters; but inputcount returns only 1 as the link to the condition is the only explicit parameter;
         //links to operations are implicit. Needed for saving though so we overrule here
@@ -228,7 +228,7 @@ void WorkflowSerializerV1::loadNode(SPWorkFlowNode& node,Workflow *workflow, con
     _stream >> parmCount;
 
     for(qint32 j=0; j < parmCount; ++j){
-        quint64 order;
+        qint32 order;
 
         _stream >> order;
         _stream >> nm;
